Adding text about non-discrimination
I think it's important to be explicit in our core values that we do not discriminate as we empower our team members with opportunities for growth. I feel this is assumed by all of us and in practice, and it should be explicitly stated somewhere in our values. I think often workplaces find excuses or put up barriers for people based on various things. Some of these things are protected classes in the United States of America, and some are common excuses for actively preventing someone from growing or being promoted within the company. I think we are a very flexible and empowering company and even though this is a negative statement (as in what we do NOT do), I think it's important to state clearly in our core values.